[#116] HistoryService 추가 및 TodoService의 로직 변경
